requirements.txt
openai
python-dotenv
gradio
PyPDF2
streamlit
azure-kusto-data
pyautogen
vision-agent
azure-cognitiveservices-speech
bs4
streamlit-audiorecorder
ffmpeg-python
pyaudio
azure-ai-translation-text==1.0.0b1
python-docx